telegram-bot/videogeneration/requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.10.1
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
coverage==7.8.0
decorator==4.4.2
emoji==2.14.1
exceptiongroup==1.2.2
frozenlist==1.6.0
gigachat==0.1.39.post1
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
loguru==0.7.2
magic-filter==1.0.12
moviepy==1.0.3
multidict==6.4.3
numpy==1.25.0
oauthlib==3.2.2
packaging==25.0
pillow==10.4.0
pluggy==1.5.0
proglog==0.1.11
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.3
pytest==8.3.5
pytest-cov==6.1.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
win32_setctime==1.2.0
yarl==1.20.0
